Moving test interceptor helper to correct location
