+++ Fitbit.Portable.Tests/Fitbit.Portable.Tests.csproj
@@ -131,7 +131,7 @@
     <Compile Include="Helper.cs" />
     <Compile Include="Helpers\ResponseFaker.cs" />
     <Compile Include="HttpClientFactoryTests.cs" />
-    <Compile Include="InterceptorCounter.cs" />
+    <Compile Include="Interceptors\InterceptorCounter.cs" />
     <Compile Include="Interceptors\FitbitHttpErrorHandlerTests.cs" />
     <Compile Include="IntradayTimeSeriesTests.cs" />
     <Compile Include="JsonDotNetSerializerTests.cs" />
